// src/generated/types.tsx
export type Maybe<T> = T | null;
// ...
export type ComplexTextValue = {
  __typename?: 'ComplexTextValue';
  /** HTML format */
  html: Scalars['String'];
};
// ...
export type PriceRange = {
  __typename?: 'PriceRange';
  /** The highest possible price for the product. */
  maximum_price?: Maybe<ProductPrice>;
  /** The lowest possible price for the product. */
  minimum_price: ProductPrice;
};
// ...
export type ProductImage = MediaGalleryInterface & {
  __typename?: 'ProductImage';
  /** Whether the image is hidden from view. */
  disabled?: Maybe<Scalars['Boolean']>;
  /** The label of the product image or video. */
  label?: Maybe<Scalars['String']>;
  /** The media item's position after it has been sorted. */
  position?: Maybe<Scalars['Int']>;
  /** The URL of the product image or video. */
  url?: Maybe<Scalars['String']>;
};
// ...
export type ProductBundleFragment = {
  __typename?: 'BundleProduct';
  uid: string;
  sku?: string | null | undefined;
  url_key?: string | null | undefined;
  url_suffix?: string | null | undefined;
  name?: string | null | undefined;
  id: string;
  description?: { __typename?: 'ComplexTextValue'; html: string } | null | undefined;
  thumbnail?:
    | {
        __typename?: 'ProductImage';
        url?: string | null | undefined;
        label?: string | null | undefined;
        id?: string | null | undefined;
      }
    | null
    | undefined;
  media_gallery?:
    | Array<
        | {
            __typename?: 'ProductImage';
            label?: string | null | undefined;
            url?: string | null | undefined;
            id?: string | null | undefined;
            type: 'ProductImage';
          }
        | {
            __typename?: 'ProductVideo';
            label?: string | null | undefined;
            url?: string | null | undefined;
            id?: string | null | undefined;
            type: 'ProductVideo';
          }
        | null
        | undefined
      >
    | null
    | undefined;
  price_range: {
    __typename?: 'PriceRange';
    minimum_price: {
      __typename?: 'ProductPrice';
      discount?:
        | { __typename?: 'ProductDiscount'; amount_off?: number | null | undefined }
        | null
        | undefined;
      regular_price: {
        __typename?: 'Money';
        currency?: CurrencyEnum | null | undefined;
        value?: number | null | undefined;
      };
    };
    maximum_price?:
      | {
          __typename?: 'ProductPrice';
          discount?:
            | {
                __typename?: 'ProductDiscount';
                amount_off?: number | null | undefined;
              }
            | null
            | undefined;
          regular_price: {
            __typename?: 'Money';
            currency?: CurrencyEnum | null | undefined;
            value?: number | null | undefined;
          };
        }
      | null
      | undefined;
  };
};

export type ProductConfigurableFragment = {
  __typename?: 'ConfigurableProduct';
  uid: string;
  sku?: string | null | undefined;
  url_key?: string | null | undefined;
  url_suffix?: string | null | undefined;
  name?: string | null | undefined;
  id: string;
  description?: { __typename?: 'ComplexTextValue'; html: string } | null | undefined;
  thumbnail?:
    | {
        __typename?: 'ProductImage';
        url?: string | null | undefined;
        label?: string | null | undefined;
        id?: string | null | undefined;
      }
    | null
    | undefined;
  media_gallery?:
    | Array<
        | {
            __typename?: 'ProductImage';
            label?: string | null | undefined;
            url?: string | null | undefined;
            id?: string | null | undefined;
            type: 'ProductImage';
          }
        | {
            __typename?: 'ProductVideo';
            label?: string | null | undefined;
            url?: string | null | undefined;
            id?: string | null | undefined;
            type: 'ProductVideo';
          }
        | null
        | undefined
      >
    | null
    | undefined;
  price_range: {
    __typename?: 'PriceRange';
    minimum_price: {
      __typename?: 'ProductPrice';
      discount?:
        | { __typename?: 'ProductDiscount'; amount_off?: number | null | undefined }
        | null
        | undefined;
      regular_price: {
        __typename?: 'Money';
        currency?: CurrencyEnum | null | undefined;
        value?: number | null | undefined;
      };
    };
    maximum_price?:
      | {
          __typename?: 'ProductPrice';
          discount?:
            | {
                __typename?: 'ProductDiscount';
                amount_off?: number | null | undefined;
              }
            | null
            | undefined;
          regular_price: {
            __typename?: 'Money';
            currency?: CurrencyEnum | null | undefined;
            value?: number | null | undefined;
          };
        }
      | null
      | undefined;
  };
  configurable_options?:
    | Array<
        | {
            __typename?: 'ConfigurableProductOptions';
            uid: string;
            label?: string | null | undefined;
            position?: number | null | undefined;
            use_default?: boolean | null | undefined;
            attribute_code?: string | null | undefined;
            product_id?: number | null | undefined;
            id: string;
            values?:
              | Array<
                  | {
                      __typename?: 'ConfigurableProductOptionsValues';
                      uid?: string | null | undefined;
                      label?: string | null | undefined;
                      id?: string | null | undefined;
                    }
                  | null
                  | undefined
                >
              | null
              | undefined;
          }
        | null
        | undefined
      >
    | null
    | undefined;
};

export type ProductDownloadableFragment = {
  __typename?: 'DownloadableProduct';
  uid: string;
  sku?: string | null | undefined;
  url_key?: string | null | undefined;
  url_suffix?: string | null | undefined;
  name?: string | null | undefined;
  id: string;
  description?: { __typename?: 'ComplexTextValue'; html: string } | null | undefined;
  thumbnail?:
    | {
        __typename?: 'ProductImage';
        url?: string | null | undefined;
        label?: string | null | undefined;
        id?: string | null | undefined;
      }
    | null
    | undefined;
  media_gallery?:
    | Array<
        | {
            __typename?: 'ProductImage';
            label?: string | null | undefined;
            url?: string | null | undefined;
            id?: string | null | undefined;
            type: 'ProductImage';
          }
        | {
            __typename?: 'ProductVideo';
            label?: string | null | undefined;
            url?: string | null | undefined;
            id?: string | null | undefined;
            type: 'ProductVideo';
          }
        | null
        | undefined
      >
    | null
    | undefined;
  price_range: {
    __typename?: 'PriceRange';
    minimum_price: {
      __typename?: 'ProductPrice';
      discount?:
        | { __typename?: 'ProductDiscount'; amount_off?: number | null | undefined }
        | null
        | undefined;
      regular_price: {
        __typename?: 'Money';
        currency?: CurrencyEnum | null | undefined;
        value?: number | null | undefined;
      };
    };
    maximum_price?:
      | {
          __typename?: 'ProductPrice';
          discount?:
            | {
                __typename?: 'ProductDiscount';
                amount_off?: number | null | undefined;
              }
            | null
            | undefined;
          regular_price: {
            __typename?: 'Money';
            currency?: CurrencyEnum | null | undefined;
            value?: number | null | undefined;
          };
        }
      | null
      | undefined;
  };
};

export type ProductGroupedFragment = {
  __typename?: 'GroupedProduct';
  uid: string;
  sku?: string | null | undefined;
  url_key?: string | null | undefined;
  url_suffix?: string | null | undefined;
  name?: string | null | undefined;
  id: string;
  description?: { __typename?: 'ComplexTextValue'; html: string } | null | undefined;
  thumbnail?:
    | {
        __typename?: 'ProductImage';
        url?: string | null | undefined;
        label?: string | null | undefined;
        id?: string | null | undefined;
      }
    | null
    | undefined;
  media_gallery?:
    | Array<
        | {
            __typename?: 'ProductImage';
            label?: string | null | undefined;
            url?: string | null | undefined;
            id?: string | null | undefined;
            type: 'ProductImage';
          }
        | {
            __typename?: 'ProductVideo';
            label?: string | null | undefined;
            url?: string | null | undefined;
            id?: string | null | undefined;
            type: 'ProductVideo';
          }
        | null
        | undefined
      >
    | null
    | undefined;
  price_range: {
    __typename?: 'PriceRange';
    minimum_price: {
      __typename?: 'ProductPrice';
      discount?:
        | { __typename?: 'ProductDiscount'; amount_off?: number | null | undefined }
        | null
        | undefined;
      regular_price: {
        __typename?: 'Money';
        currency?: CurrencyEnum | null | undefined;
        value?: number | null | undefined;
      };
    };
    maximum_price?:
      | {
          __typename?: 'ProductPrice';
          discount?:
            | {
                __typename?: 'ProductDiscount';
                amount_off?: number | null | undefined;
              }
            | null
            | undefined;
          regular_price: {
            __typename?: 'Money';
            currency?: CurrencyEnum | null | undefined;
            value?: number | null | undefined;
          };
        }
      | null
      | undefined;
  };
};

export type ProductSimpleFragment = {
  __typename?: 'SimpleProduct';
  uid: string;
  sku?: string | null | undefined;
  url_key?: string | null | undefined;
  url_suffix?: string | null | undefined;
  name?: string | null | undefined;
  id: string;
  description?: { __typename?: 'ComplexTextValue'; html: string } | null | undefined;
  thumbnail?:
    | {
        __typename?: 'ProductImage';
        url?: string | null | undefined;
        label?: string | null | undefined;
        id?: string | null | undefined;
      }
    | null
    | undefined;
  media_gallery?:
    | Array<
        | {
            __typename?: 'ProductImage';
            label?: string | null | undefined;
            url?: string | null | undefined;
            id?: string | null | undefined;
            type: 'ProductImage';
          }
        | {
            __typename?: 'ProductVideo';
            label?: string | null | undefined;
            url?: string | null | undefined;
            id?: string | null | undefined;
            type: 'ProductVideo';
          }
        | null
        | undefined
      >
    | null
    | undefined;
  price_range: {
    __typename?: 'PriceRange';
    minimum_price: {
      __typename?: 'ProductPrice';
      discount?:
        | { __typename?: 'ProductDiscount'; amount_off?: number | null | undefined }
        | null
        | undefined;
      regular_price: {
        __typename?: 'Money';
        currency?: CurrencyEnum | null | undefined;
        value?: number | null | undefined;
      };
    };
    maximum_price?:
      | {
          __typename?: 'ProductPrice';
          discount?:
            | {
                __typename?: 'ProductDiscount';
                amount_off?: number | null | undefined;
              }
            | null
            | undefined;
          regular_price: {
            __typename?: 'Money';
            currency?: CurrencyEnum | null | undefined;
            value?: number | null | undefined;
          };
        }
      | null
      | undefined;
  };
};